+++ packages/patternfly-4/react-core/src/components/Switch/index.d.ts
@@ -1,0 +1,1 @@
+export { default as Switch, SwitchProps } from './Switch';